% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no-export-fns.R
\name{char2FUN}
\alias{char2FUN}
\title{Convert JAGS Model from Character to Function Format}
\usage{
char2FUN(char)
}
\arguments{
\item{char}{Character vector with elements storing distinct lines of JAGS
model code}
}
\value{
Function containing JAGS model code
}
\description{
Convert JAGS Model from Character to Function Format
}
\note{
For internal use only, users need not concern
themselves with this function
}
