% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no-export-fns.R
\name{check_FUN}
\alias{check_FUN}
\title{Check \code{FUN} Argument}
\usage{
check_FUN(FUN)
}
\arguments{
\item{FUN}{Function containing JAGS model code.}
}
\description{
Standardized check for function
}
\details{
Simply checks if \code{is.function(FUN)}, and if \code{FALSE},
returns an error
}
\note{
For internal use only, users need not concern
themselves with this function
}
