% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no-export-fns.R
\name{check_model_file}
\alias{check_model_file}
\title{Check \code{model_file} Argument}
\usage{
check_model_file(model_file = NULL)
}
\arguments{
\item{model_file}{Name of file containing JAGS model code}
}
\description{
Standardized check for JAGS model file names
}
\details{
Currently simply checks if is \code{NULL},
but more informative checks will likely be added in future versions
}
