% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no-export-fns.R
\name{drop_lws}
\alias{drop_lws}
\title{Drop Leading Whitespace}
\usage{
drop_lws(x, n = NULL)
}
\arguments{
\item{x}{Character string that has leading white space to be removed}

\item{n}{number of spaces to remove, default is \code{NULL} to remove all}
}
\description{
Shortcut for \code{stringr::str_remove(x, "^\\\\s+")}
}
\note{
For internal use only, users need not concern
themselves with this function
}
