% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no-export-fns.R
\name{FUN2char}
\alias{FUN2char}
\title{Convert JAGS Model from Function to Character Format}
\usage{
FUN2char(FUN, keep_comments = TRUE)
}
\arguments{
\item{FUN}{Function containing JAGS model code}

\item{keep_comments}{Logical flag indicating whether to retain comments
(i.e., lines starting with \verb{#}) in the output (default is \code{TRUE})}
}
\value{
Character vector with elements containing distinct lines of code from
the body of the function supplied to \code{FUN}
}
\description{
Isolates the function body into a standard
character vector
}
\note{
For internal use only, users need not concern
themselves with this function
}
