% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no-export-fns.R
\name{jags_styler}
\alias{jags_styler}
\title{Style code for JAGS}
\usage{
jags_styler(x)
}
\arguments{
\item{x}{Character vector with elements storing distinct lines of JAGS
model code}
}
\description{
Ensures printed code is formatted nicely and easily readable by humans
}
\details{
Primarily used for managing indentation when multiline
code is wrapped in \code{{}}
}
\note{
For internal use only, users need not concern themselves
with this function
}
