% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{jagsMB_opts}
\alias{jagsMB_opts}
\title{Set or Get jagsMB Options}
\usage{
jagsMB_opts(...)
}
\arguments{
\item{...}{Option names to retrieve options values or \code{key = value} pairs
to set options. The only value currently accepted is \code{model_file},
which serves as the default argument value for \code{\link[=model_read]{model_read()}},
\code{\link[=model_lines]{model_lines()}}, \code{\link[=model_vars]{model_vars()}}, and \code{\link[=model_search]{model_search()}}.}
}
\description{
Set or Get jagsMB Options
}
