% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-fns.R
\name{model_header}
\alias{model_header}
\title{Create a Section Header to Insert into JAGS Model}
\usage{
model_header(text = "text", decor = "### --- ", toupper = TRUE)
}
\arguments{
\item{text}{Character string with the text to insert in the
header block}

\item{decor}{Character string containing the "decoration" that
will precede the text in the header. Will be reversed and placed at end of
header text as well (default is \code{"### --- "}).}

\item{toupper}{Logical flag indicating whether to convert all letters to
upper case (default is \code{TRUE}).}
}
\value{
Function ready to be included as a list element
and passed to \code{\link[=model_build]{model_build()}}.
}
\description{
Constructs a section header to be placed in a JAGS model file.
}
\note{
If the first character of \code{decor} is not \verb{#} or if either \code{text} or
\code{decor} have length > 1, an error will be returned.
}
