% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-fns.R
\name{model_read}
\alias{model_read}
\title{Read Model File}
\usage{
model_read(model_file = jagsMB_opts("model_file"), keep_lws = TRUE)
}
\arguments{
\item{model_file}{Name of file containing JAGS model code;
defaults to \code{jagsMB_opts("model_file")}.}

\item{keep_lws}{Logical flag indicating whether to retain leading white space
on each line (default is \code{TRUE}).}
}
\value{
A character vector with elements storing each line of the model file
contents.
}
\description{
Wrapper for \code{\link[base:readLines]{base::readLines()}} but with defaults and
styling to handle indentation
}
