% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-fns.R
\name{model_search}
\alias{model_search}
\title{Search Model File for a Pattern}
\usage{
model_search(model_file = jagsMB_opts("model_file"), pattern, side = "any")
}
\arguments{
\item{model_file}{Name of file containing JAGS model code;
defaults to \code{jagsMB_opts("model_file")}.}

\item{pattern}{Character string to find matches for in the model
definition found in \code{model_file}. Supplied to \code{\link[stringr:str_which]{stringr::str_which()}}
so may be regex.}

\item{side}{Character string; one of \code{"left"}, \code{"right"}, or \code{"any"};
the side of the assignment operator (\code{~} or \verb{<-} the string matching
\code{pattern} is found).}
}
\description{
Returns the line numbers containing a
pattern match, with the option to search
on either the left- or right-hand side
of the assignment operator on that line.
}
