% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-fns.R
\name{model_vars}
\alias{model_vars}
\title{List all Variable Names in JAGS Model}
\usage{
model_vars(model_file = jagsMB_opts("model_file"))
}
\arguments{
\item{model_file}{Name of file containing JAGS model code;
defaults to \code{jagsMB_opts("model_file")}.}
}
\value{
A list object with elements storing the types of variables
created within the JAGS model:
\itemize{
\item \verb{$logical} variables defined to the left of a \verb{<-} operator.
\item \verb{$stochastic}: variables defined to the left of a \code{~} operator.
}
}
\description{
Returns the names of all variables defined
within the JAGS model code
(i.e., found on the left-hand side of \verb{<-} or \code{~}).
}
