% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-fns.R
\name{model_write}
\alias{model_write}
\title{Write JAGS Model Definition File}
\source{
This function was heavily influenced by \code{\link[R2OpenBUGS:write.model]{R2OpenBUGS::write.model()}}.
}
\usage{
model_write(FUN, model_file = NULL, keep_comments = TRUE, keep_empty = TRUE)
}
\arguments{
\item{FUN}{Function containing JAGS model code.}

\item{model_file}{File name to write model code to;
must have \code{".txt"} extension.
Default (\code{NULL}) is to create a temporary file name with
\code{tempfile(pattern = "", fileext = ".txt")}.}

\item{keep_comments}{Logical flag indicating whether to retain comments
(i.e., lines starting with \verb{#}) in the output (default \code{TRUE}).}

\item{keep_empty}{Logical flag indicating whether to retain entirely
empty lines.}
}
\value{
File name of the written model (returned invisibly); same value
as \code{model_file}.
}
\description{
Wrapper for \code{\link[base:writeLines]{base::writeLines()}} but with defaults and
styling to handle indentation
}
\note{
\code{model_write()} and \code{\link[=model_replace]{model_replace()}} are the only functions that
accept a \code{model_file} argument but do not default to
\verb{jagsMB_opts("model_file)"}. Instead they default to \code{NULL}, which returns
an error requiring the user to supply a file path manually.
}
